requirements.txt
polars>=0.20.0
pytest
PyYAML